Move flatten spreadsheet into convert
